requirements.txt
fastapi
uvicorn
python-digitalocean